fastapi_app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
mlflow==2.13.0
pydantic==2.7.1
azure-storage-blob==12.19.1
